Add files via upload
I was playing a game that uses resin to do bosses and domains and the resin counter goes up by 1 every 8 minutes. I was getting annoyed having to pull up a calculator every single time I wanted to guess approximately when I could log in to use the resin so I made a resin calculator instead.